fixed bugs time to work on making map

+++ server/controllers/users.js
@@ -4,6 +4,10 @@ import fs from "fs";
 
 export const patchSettings = async (req, res) => {
     try{
+        if(req.fileExt && ![".jpg", ".png", ".jpeg"].includes(req.fileExt)){
+            throw new Error("Invalid file format");
+        }
+
         const {userId} = req.params;
         const fields = ['firstName', 'lastName', 'username', 'email', 'password'];
         const updates = Object.fromEntries(Object.entries(req.body)
@@ -19,10 +23,23 @@ export const patchSettings = async (req, res) => {
             const passwordHash = await bcrypt.hash(updates.password, salt);
             updates.password = passwordHash;
         }
-        const updatedUser = await User.findByIdAndUpdate(userId, updates, {new: true});
+
+        //returns user before the update
+        const oldUser = await User.findByIdAndUpdate(userId, updates);
+        if(!oldUser) return res.status(401).json({error: "User does not exist."})
+            
+        //if changed the user's picture path (by using different file format), delete old one
+        if(req.dbPicturePath && oldUser.picturePath !== req.dbPicturePath){
+            fs.unlink(oldUser.picturePath, (err) => {
+                if(err) return console.log(err);
+                console.log(`${oldUser.picturePath} was deleted`);
+            }); 
+        }
+
         if(req.curPicturePath){
+            //rename the pending picture to the actual pfp
             fs.rename(req.curPicturePath, req.dbPicturePath, (err) => {
-                //this shouldn't ever error, but if it did it would just result in the user having no pfp
+                //this shouldn't ever error, but if it did it would just result in the user having old pfp (or no pfp)
                 //which they could change in settings
                 if(err){
                     console.log(err);
@@ -35,6 +52,7 @@ export const patchSettings = async (req, res) => {
                 console.log(`Renamed ${req.curPicturePath} to ${req.dbPicturePath}`);
             });
         }
+        const updatedUser = await User.findById(userId);
         return res.status(200).json({user: updatedUser});
     } catch(err){
         if(req.curPicturePath){
